+++ test/generator/filler/test_runner.py
@@ -118,9 +118,7 @@ def _make_pickup(item_category: ItemCategory):
         model_index=0,
         item_category=item_category,
         broad_category=item_category,
-        resources=(
-            ConditionalResources(None, None, ()),
-        ),
+        progression=tuple(),
     )